requirements.txt
django
djangorestframework
django-cors-headers
mysqlclient
whitenoise
django-celery-beat

scrapy
mariadb

celery
redis

python-dotenv
gunicorn
beautifulsoup4

pandas
numpy
